week 2 loppuun

+++ database/selectdata.php
@@ -1,6 +1,12 @@
 <?php
+session_start();
+if (!isset($_SESSION['user'])) {
+    header('Location: index.php');
+    exit;
+}
+
 global $DBH;
-require_once 'dbconnect.php';
+require_once 'dbConnect.php';
 
 $sql = 'SELECT * FROM MediaItems;';
 
@@ -17,13 +23,18 @@ try {
         echo '<td>' . $row['title'] . '</td>';
         echo '<td>' . $row['description'] . '</td>';
         echo '<td>' . $row['created_at'] . '</td>';
-        echo '<td>
+        if ($_SESSION['user']['user_id'] == $row['user_id']) {
+            echo '<td>
                 <a href="deleteData.php?id=' . $row['media_id'] . '">Delete</a>
                 <a href="#" class="modify-link" data-id="' . $row['media_id'] . '">Modify</a>   
               </td>';
+        } else {
+            echo '<td>Ei kuulu sulle</td>';
+        }
         echo '</tr>';
     }
 } catch (PDOException $e) {
     echo "Could not select data from the database.";
     file_put_contents('PDOErrors.txt', 'selectData.php - ' . $e->getMessage(), FILE_APPEND);
 }
+
